add more structure
