--- src/NexGrades/NexGrades.App/Pages/StudentPage.xaml.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using NexGrades.App.ViewModels;
-using Wpf.Ui.Abstractions.Controls;
-
-namespace NexGrades.App.Pages
-{
-    /// <summary>
-    /// Interaction logic for StudentPage.xaml
-    /// </summary>
-    public partial class StudentPage : INavigableView<StudentViewModel>
-    {
-        public StudentPage(StudentViewModel vm)
-        {
-            ViewModel = vm;
-            DataContext = this;
-
-            InitializeComponent();
-        }
-
-        public StudentViewModel ViewModel { get; }
-    }
-}
