+++ src/NexGrades/NexGrades.App/Pages/Students/StudentsOverviewPage.xaml
@@ -1,4 +1,4 @@
-﻿<Page x:Class="NexGrades.App.Pages.StudentsOverviewPage"
+﻿<Page x:Class="NexGrades.App.Pages.Students.StudentsOverviewPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
