+++ src/NexGrades/NexGrades.App/Pages/Students/StudentsOverviewPage.xaml.cs
@@ -1,21 +1,7 @@
 ﻿using NexGrades.App.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using Wpf.Ui.Abstractions.Controls;
 
-namespace NexGrades.App.Pages
+namespace NexGrades.App.Pages.Students
 {
     /// <summary>
     /// Interaction logic for StudentsOverviewPage.xaml
